Rename utils to nodes and fix circular import
- move containers to associated files
