+++ tests/test_four_c_simulation.py
@@ -50,7 +50,7 @@ from meshpy.four_c.dbc_monitor import (
 from meshpy.four_c.run_four_c import run_four_c
 from meshpy.mesh_creation_functions.beam_basic_geometry import create_beam_mesh_line
 from meshpy.mesh_creation_functions.beam_honeycomb import create_beam_mesh_honeycomb
-from meshpy.utils.utils import check_node_by_coordinate
+from meshpy.utils.nodes import check_node_by_coordinate
 
 # We test all test cases in this file twice. Once we only run up to the first
 # call of 4C and compare the created input files, this allows to run some core
